backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
scipy==1.12.0
python-multipart==0.0.6
rasterio==1.4.3

streamlit
openai==0.28
audio_recorder_streamlit
streamlit-float
python-dotenv
assemblyai
pyttsx3
google-cloud-texttospeech
requests
pydub
numpy